이미지 업로드

+++ java/com/ubi/erp/pers/controller/PersDataSController.java
@@ -1,6 +1,9 @@
 package com.ubi.erp.pers.controller;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -11,17 +14,20 @@ import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
 
+import com.ubi.erp.cmm.exception.UbiBizException;
 import com.ubi.erp.cmm.file.AttachFileService;
 import com.ubi.erp.cmm.util.PropertyUtil;
 import com.ubi.erp.cmm.util.gson.DateFormatUtil;
 import com.ubi.erp.pers.domain.PersDataS;
 import com.ubi.erp.pers.service.PersDataSService;
-import com.ubi.erp.user.domain.AttachFile;
 
 @RestController
 @RequestMapping(value = "/erp/pers/pers/persDataS")
@@ -32,6 +38,10 @@ public class PersDataSController {
 	
 	@Autowired
 	private AttachFileService attachFileService;
+	
+	private String saveFilename;
+	
+	private String filePath = "/images/temp";
 	
 	@SuppressWarnings("unchecked")
 	@RequestMapping(value = "/selLeft",method = RequestMethod.POST)
@@ -92,7 +102,10 @@ public class PersDataSController {
 		persDataS.setRetireMidDate(df.dateToString(persDataS.getRetireMidDate()));
 		// armyJong check -- armyJong == armyKind
 		persDataS.setArmyJong(persDataS.getArmyKind());
-
+		System.out.println("saveFilename======"+saveFilename);
+		if(saveFilename != null){
+			persDataS.setImgPath(saveFilename);
+		}
 		if("INSERT".equals(persDataS.getCudKey())) {
 			persDataSService.prcsPersDataS(persDataS);
 		}else if("UPDATE".equals(persDataS.getCudKey())){
@@ -108,24 +121,58 @@ public class PersDataSController {
 	@RequestMapping(value = "/prcsFileUpload",method= RequestMethod.POST)
 	@ResponseStatus(HttpStatus.OK)
 	public void prcsfileUpload(HttpServletRequest request, HttpServletResponse response) throws Exception {
-		List<AttachFile> uploadFileList = attachFileService.uploadAttachFile(PropertyUtil.getString("attach.savedir"), request, response);
-		PersDataS persDataS = new PersDataS();
-		persDataS.setCompId(request.getParameter("imgCompId"));
-		persDataS.setEmpNo(request.getParameter("imgEmpNo"));
-		persDataS.setCompId("100");
-		persDataS.setEmpNo("1230981");
-		persDataS.setImgPath(uploadFileList.get(0).getFilePath());
-		persDataSService.updateImgPath(persDataS);
+		String saveDir = request.getSession().getServletContext().getRealPath(filePath);
+
+		MultipartHttpServletRequest multipartRequest = (MultipartHttpServletRequest) request;
+		MultiValueMap<String, MultipartFile> multiValueMap = multipartRequest.getMultiFileMap();
+
+		List<MultipartFile> files = multiValueMap.get("fileName");
+		for (MultipartFile file : files) {
+			if (!file.isEmpty()) {
+				long limitSize = Long.parseLong(PropertyUtil.getString("attach.uploadSize"));
+				
+				if (limitSize > file.getSize()) {
+					String fileName = file.getOriginalFilename();
+					String ext = fileName.substring(fileName.lastIndexOf(".") + 1);
+					String onlyName = fileName.substring(0, fileName.lastIndexOf("."));
+					saveFilename = fileName;
+			
+					// 파일명이 중복되는 경우 변경처리
+					if (new File(saveDir + "/" + fileName).exists()) {
+						int fileSeq = 1;
+						while (isFileExists(saveDir, onlyName, fileSeq, ext)) {
+							fileSeq++;
+						}
+						saveFilename = onlyName + "_" + fileSeq + "." + ext;
+					}
+			
+					// 실제 파일 업로드
+					file.transferTo(new File(saveDir + "/" + saveFilename));
+
+				} 
+			}
+		}
 	}
 	
 	@RequestMapping(value = "/prcsFileDelete",method= RequestMethod.POST)
 	@ResponseStatus(HttpStatus.OK)
-	public void prcsfileDelete(HttpServletRequest request, HttpServletResponse response,PersDataS persDataS) throws Exception {	
-		File targetFile = new File(PropertyUtil.getString("attach.basedir") + persDataS.getImgPath());
-		targetFile.delete();
-		//persDataSService.prcsPersDataS(persDataS);
+	public void prcsfileDelete(HttpServletRequest request, HttpServletResponse response,HttpSession session,PersDataS persDataS) throws Exception {	
+		String delDir = request.getSession().getServletContext().getRealPath(filePath);
+		File targetFile = new File(delDir,persDataS.getImgPath());
+		if (targetFile.exists()) {
+			targetFile.delete();
+		}
+		persDataS.setImgPath("");
+		prcsPersData(request, response, session, persDataS);
 	}
 	
+	public static void fileDelete(String upDir, String oldfile) {
+		File file = new File(upDir, oldfile);
+
+		if (file.exists()) {
+			file.delete();
+		}
+	}
 	
 	public String nullCheck(String value){
 		if(value == null){
@@ -133,4 +180,9 @@ public class PersDataSController {
 		}
 		return value;
 	}
+	
+	public boolean isFileExists(String saveDir, String onlyName, int fileSeq, String ext) {
+		return new File(saveDir + "/" + onlyName + "_" + (fileSeq) + "." + ext).exists();
+	}
+	
 }
